refractor: move component for controls into their own component
