+++ src/components/React/volume-viewer/VolumeViewer.js
@@ -2,7 +2,7 @@ import { Container } from "react-bootstrap";
 import { Sidebar } from "primereact/sidebar";
 
 import VolumeRenderer from "./VolumeRenderer";
-import Controls from "../Controls"
+import Controls from "../controls/Controls"
 
 export default function VolumeViewer(props) {
   const {sidebarVisible, setSidebarVisible } = props
